Move postgres*_version_to_int static methods to own file
